Reduced many types to float32 to save memory moved maths functions to math.go implemented flat HiddenNeurons array without pointers to reduce memmove consolidated meat and food into single food cache fixed text rendering

+++ v2/world/coord.go
@@ -4,15 +4,15 @@ import "math"
 
 // Position is a continuous 2D coordinate in world-space.
 type Position struct {
-	X, Y float64
+	X, Y float32
 }
 
-func (p Position) DistanceTo(other Position) float64 {
+func (p Position) DistanceTo(other Position) float32 {
 	dx := p.X - other.X
 	dy := p.Y - other.Y
-	return math.Sqrt(dx*dx + dy*dy)
+	return float32(math.Sqrt(float64(dx*dx + dy*dy)))
 }
 
-func (p Position) Add(dx, dy float64) Position {
+func (p Position) Add(dx, dy float32) Position {
 	return Position{p.X + dx, p.Y + dy}
 }
